requirements.txt
requests
beautifulsoup4
pysocks
openai
pandas
python-dotenv
playwright
pytesseract
Pillow
